Replace gps.inputOnly config file property by gps.config
Part of #12.
Refactor GPS Configure functions to take a gpsprot.ConfigTarget struct,
which combines existing ConfigMap and ConfigOptions with
the CfgKeySet specifying the set of keys to be got from the GPS.
gps.config=false is represented by a ConfigTarget that has an empty map and no ConfigOptions that change anything.
Make the UBX legacy config functions only poll for the properties
that are needed.