Use unix package implementation of PTP ioctls
These got added recently, and covers almost all of what was in our unix2 package.
The only remaining thing was SerialICounter used by term,
so I moved that into the term package.
This allows removal of unix2 package.
The newly added interface is similar, but not identical to what I had, so there are some changes.