Accept NMEA-like packets
The NMEA PacketFormat does not recognize some packets that are NMEA-like but properly NMEA conformant.
This is a problem in implementing support for Unicore receivers, which use these kinds of packets.
These are used by older generation Unicore receivers extensively and
also by newer receivers (such as UM980) in some situations (e.g. responses to configuration commands).
We solve this problem by making the PacketFormat accept these packets,
The NMEA-validation that was done previously by the PacketFormat is now done by the PacketProcessor.
The change includes a design document nmea-lax.md.
This also fixes handling of NMEA proprietary sentences (starting with P)
as described in issue #74.