Make scan handle packet formats with common prefixes
Scanner.Scan currently can't handle having packet formats which  can start with the same character.
We don't have this currently.
But Unicore and NovAtel OEM7 binary formats have 3 byte sync sequences that start with the same two bytes.
This hasn't been tested yet, so it's just a draft to record the idea for solving it that I had.