Avoid filesystem paths for local end of Unix sockets
This is talking to chrony and ptp4l.
Previously we created filesystem paths for the local end.
This caused problems with AppArmor and chrony on Ubuntu 24.
Instead use an anonymous socket with chrony, which doesn't need replies.
Use abstract namespace for the local end with ptp4l, which does need to reply.
This significantly simplifies the code, because there is no need to remove filesystem inodes.
Fixes #165.